src/pyproject.toml
[project]
name = "llm-pipeline"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.40.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "jsonschema>=4.21.0",
    "tqdm>=4.66.0",
]

[project.scripts]
llm-pipeline = "llm_pipeline.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llm_pipeline"]

